feat(philosophers): fixed the bug of hitting eaten number but not terminated. more of less done, need norminate

+++ philo/src/cleanup/free_table.c
@@ -6,7 +6,7 @@
 /*   By: ylai <ylai@student.42singapore.sg>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/31 18:10:51 by ylai              #+#    #+#             */
-/*   Updated: 2024/12/07 18:16:01 by ylai             ###   ########.fr       */
+/*   Updated: 2024/12/14 17:59:44 by ylai             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ void	free_forks(t_table *table, int fr_err)
 	int	i;
 
 	i = 0;
-	while (fr_err == -1 && (unsigned long)i < table->philo_num)
+	while (fr_err == -1 && i < table->philo_num)
 	{
 		pthread_mutex_destroy(&(table->forks[i]));
 		i++;
@@ -27,7 +27,6 @@ void	free_forks(t_table *table, int fr_err)
 		pthread_mutex_destroy(&(table->forks[i]));
 		i++;
 	}
-	// free(table->forks);
 }
 
 void	free_mut(t_table *table, int mu_err)
@@ -53,5 +52,4 @@ void	free_table(t_table *table, int mu_err, int tr_err, int fr_err)
 	free_mut(table, mu_err);
 	// free_philo(table, tr_err);
 	(void) tr_err;
-	// free(table->philo);
 }
